requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pymupdf==1.23.8
torch==2.1.2
transformers==4.36.2
pillow==10.2.0
numpy==1.26.3
scikit-image==0.22.0
jinja2==3.1.3
aiofiles==23.2.1
